refactor: split gcode import models
